[MNG-5594] Output current/total project number
This adds a progress indicator to the output of a multi-module reactor
build:

- add [N] numbers before projects in "Reactor Build Order" at start
- add [N/M] progress before every "Building ..." message
- add [N] numbers before projects in "Reactor Summary" at end

Progress numbers refer to the project numbers printed at
start. Therefore, in parallel builds the output may not be strictly
ordered as in sequential builds. However, it may help to understand
the true execution ordering in a parallel build.